requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
astroid==2.9.3
beir==0.0.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.2
django-allauth==0.49.0
django-crispy-forms==1.14.0
easy-elasticsearch==0.0.7
elasticsearch==7.12.1
filelock==3.6.0
gevent==21.12.0
gql==3.1.0
graphql-core==3.2.0
greenlet==1.1.2
grequests==0.6.0
huggingface-hub==0.4.0
idna==3.3
isort==5.10.1
joblib==1.1.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
multidict==6.0.2
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
promise==2.3
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
pyparsing==3.0.7
python3-openid==3.2.0
pytrec-eval==0.5
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
Rx==1.6.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
threadpoolctl==3.1.0
tokenizers==0.11.6
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
transformers==4.17.0
typing_extensions==4.1.1
urllib3==1.26.8
wrapt==1.13.3
yarl==1.7.2
yelpapi==2.4.0
zope.event==4.5.0
zope.interface==5.4.0
